+++ src/gam/settings/_gam.py
@@ -38,5 +38,3 @@ DISCORD_COGS = [
 DISCORD_KEY = getenv("DISCORD_KEY", None)
 
 DISCORD_COMMAND_PREFIX = "!"
-
-GAM_COINS_PER_MINUTE = 10
